+++ Project 1/src_files/i2c.c
@@ -136,7 +136,7 @@ int I2C_Write_Word(int file , uint8_t *data)
 }
 
 //Write I2C data with 16bit data
-int I2C_Write_Bytes(int file , uint8_t * data)
+int I2C_Write_Bytes(int file , uint16_t * data)
 {
 	if(write(file,data,2)!=2)
 	{
